+++ src/main/java/com/perimeterx/http/IPXHttpClient.java
@@ -1,0 +1,10 @@
+package com.perimeterx.http;
+
+import java.io.Closeable;
+import java.io.IOException;
+
+public interface IPXHttpClient extends Closeable {
+    IPXIncomingResponse send(IPXOutgoingRequest request) throws IOException;
+
+    void sendAsync(IPXOutgoingRequest request) throws IOException;
+}