+++ src/main/java/com/perimeterx/http/IPXIncomingResponse.java
@@ -1,0 +1,11 @@
+package com.perimeterx.http;
+
+import java.io.Closeable;
+import java.io.IOException;
+import java.io.InputStream;
+
+public interface IPXIncomingResponse extends Closeable {
+    InputStream body() throws IOException;
+    PXHttpStatus status();
+    PXHttpHeader[] headers();
+}