+++ src/main/java/com/perimeterx/http/PXHttpMethod.java
@@ -1,0 +1,11 @@
+package com.perimeterx.http;
+
+public enum PXHttpMethod {
+    GET,
+    POST,
+    PUT,
+    PATCH,
+    DELETE,
+    HEAD,
+    OPTIONS
+}